data_extractor/code/rule_based_pipeline/rule_based_pipeline/requirements.txt
numpy
jsonpickle
scipy
Pillow
pandas